chore: rename errors package
