+++ pkg/errutil/must.go
@@ -1,4 +1,4 @@
-package errors
+package errutil
 
 import "fmt"
